Extract PaidAdsSectionsGroup as a standalone component PaidAdsSetupSections from SetupPaidAds.
Address
- https://github.com/woocommerce/google-listings-and-ads/pull/1677#discussion_r974388734
- https://github.com/woocommerce/google-listings-and-ads/pull/1677#discussion_r974392405
